+++ modules/util.js
@@ -1,0 +1,11 @@
+const getRandomInteger = function (min, max) {
+  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
+  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+
+  return Math.floor(result);
+};
+
+const getRandomElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+export { getRandomInteger, getRandomElement };
